requirements.txt
amqp==2.3.2
Django==2.0
django-crispy-forms==1.7.2
kombu==4.2.1
Pillow==5.2.0
pytz==2018.5
vine==1.1.4